requirements.txt
streamlit
google-generativeai
Pillow
pandas
openpyxl
sqlalchemy
bcrypt